move contig collection to a different file
